# server/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Bad bank application",
        "description": "Banking Fullstack MERN Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "host": "badbank-mern-ad70a.firebaseapp.com",
    "basePath": "/",
    "tags": [
        {
            "name": "Users",
            "description": "API for users information in the database"
        }
    ],
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],

    "paths": {
        "/api/account/all": {
            "get": {
                "tags": ["Users"],
                "summary": "Get all users information in the database",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/account/create/{name}/{email}/{password}": {
            "get": {
                "tags": ["Users"],
                "summary": "Create new user in the database",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "description": "Name of the user to be created",
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "description": "Email of the user to be created",
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "required": true,
                        "description": "Password of the user to be created",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully"
                    }
                }
            }
        },
        "/api/account/login/{email}/{password}": {
            "get": {
                "tags": ["Users"],
                "summary": "Login user with email and password",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "description": "Email of the user in the database",
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "path",
                        "required": true,
                        "description": "Password of the user in the database",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    }
                }
            }
        },
        "/api/account/logout": {
            "get": {
                "tags": ["Users"],
                "summary": "Logout the current logged in user",
                "responses": {
                    "200": {
                        "description": "Logout successfull"
                    }
                }
            }
        },
        "/account/update/{email}/{amount}": {
            "get": {
                "tags": ["Users"],
                "summary": "Update the balance of the current user",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "description": "Email of the current user",
                        "type": "string"
                    },
                    {
                        "name": "amount",
                        "in": "path",
                        "required": true,
                        "description": "Amount to add to the current user balance",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Balance updated successfully"
                    }
                }
            }
        }
    }
}